+++ server/src/main/java/org/jetbrains/bsp/bazel/server/bsp/BspIntegrationData.kt
@@ -15,5 +15,4 @@ data class BspIntegrationData(
     val traceWriter: PrintWriter?,
 ) {
     lateinit var launcher: Launcher<BuildClient>
-    lateinit var server: Server
 }