+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/ProjectSyncService.java
@@ -33,107 +33,121 @@ import ch.epfl.scala.bsp4j.WorkspaceBuildTargetsResult;
 
 import java.util.Collections;
 
+import org.eclipse.lsp4j.jsonrpc.CancelChecker;
 import org.jetbrains.bsp.bazel.server.sync.model.Language;
 
 /**
  * A facade for all project sync related methods
  */
 public class ProjectSyncService {
-    private final BspProjectMapper bspMapper;
-    private final ProjectProvider projectProvider;
+  private final BspProjectMapper bspMapper;
+  private final ProjectProvider projectProvider;
 
-    public ProjectSyncService(BspProjectMapper bspProjectMapper, ProjectProvider projectProvider) {
-        this.bspMapper = bspProjectMapper;
-        this.projectProvider = projectProvider;
-    }
+  public ProjectSyncService(BspProjectMapper bspProjectMapper, ProjectProvider projectProvider) {
+    this.bspMapper = bspProjectMapper;
+    this.projectProvider = projectProvider;
+  }
 
-    public InitializeBuildResult initialize() {
-        return bspMapper.initializeServer(Language.Companion.all());
-    }
+  public InitializeBuildResult initialize(CancelChecker cancelChecker) {
+    return bspMapper.initializeServer(Language.Companion.all());
+  }
 
-    // We might consider doing the actual project reload in this endpoint
-    // i.e. just run projectProvider.refreshAndGet() and in workspaceBuildTargets
-    // just run projectProvider.get() although current approach seems to work
-    // correctly, so I am not changing anything.
-    public Object workspaceReload() {
-        return new Object();
-    }
+  // We might consider doing the actual project reload in this endpoint
+  // i.e. just run projectProvider.refreshAndGet() and in workspaceBuildTargets
+  // just run projectProvider.get() although current approach seems to work
+  // correctly, so I am not changing anything.
+  public Object workspaceReload(CancelChecker cancelChecker) {
+    return new Object();
+  }
 
-    public WorkspaceBuildTargetsResult workspaceBuildTargets() {
-        var project = projectProvider.refreshAndGet();
-        return bspMapper.workspaceTargets(project);
-    }
+  public WorkspaceBuildTargetsResult workspaceBuildTargets(CancelChecker cancelChecker) {
+    var project = projectProvider.refreshAndGet(cancelChecker);
+    return bspMapper.workspaceTargets(project);
+  }
 
-    public SourcesResult buildTargetSources(SourcesParams sourcesParams) {
-        var project = projectProvider.get();
-        return bspMapper.sources(project, sourcesParams);
-    }
+  public SourcesResult buildTargetSources(
+      CancelChecker cancelChecker, SourcesParams sourcesParams) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.sources(project, sourcesParams);
+  }
 
-    public ResourcesResult buildTargetResources(ResourcesParams resourcesParams) {
-        var project = projectProvider.get();
-        return bspMapper.resources(project, resourcesParams);
-    }
+  public ResourcesResult buildTargetResources(
+      CancelChecker cancelChecker, ResourcesParams resourcesParams) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.resources(project, resourcesParams);
+  }
 
-    public InverseSourcesResult buildTargetInverseSources(InverseSourcesParams inverseSourcesParams) {
-        var project = projectProvider.get();
-        return bspMapper.inverseSources(project, inverseSourcesParams);
-    }
+  public InverseSourcesResult buildTargetInverseSources(
+      CancelChecker cancelChecker, InverseSourcesParams inverseSourcesParams) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.inverseSources(project, inverseSourcesParams);
+  }
 
-    public DependencySourcesResult buildTargetDependencySources(
-            DependencySourcesParams dependencySourcesParams) {
-        var project = projectProvider.get();
-        return bspMapper.dependencySources(project, dependencySourcesParams);
-    }
+  public DependencySourcesResult buildTargetDependencySources(
+      CancelChecker cancelChecker, DependencySourcesParams dependencySourcesParams) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.dependencySources(project, dependencySourcesParams);
+  }
 
-    public OutputPathsResult buildTargetOutputPaths(OutputPathsParams params) {
-        var project = projectProvider.get();
-        return bspMapper.outputPaths(project, params);
-    }
+  public OutputPathsResult buildTargetOutputPaths(
+      CancelChecker cancelChecker, OutputPathsParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.outputPaths(project, params);
+  }
 
-    public JvmRunEnvironmentResult jvmRunEnvironment(JvmRunEnvironmentParams params) {
-        var project = projectProvider.get();
-        return bspMapper.jvmRunEnvironment(project, params);
-    }
+  public JvmRunEnvironmentResult jvmRunEnvironment(
+      CancelChecker cancelChecker, JvmRunEnvironmentParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.jvmRunEnvironment(project, params);
+  }
 
-    public JvmTestEnvironmentResult jvmTestEnvironment(JvmTestEnvironmentParams params) {
-        var project = projectProvider.get();
-        return bspMapper.jvmTestEnvironment(project, params);
-    }
+  public JvmTestEnvironmentResult jvmTestEnvironment(
+      CancelChecker cancelChecker, JvmTestEnvironmentParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.jvmTestEnvironment(project, params);
+  }
 
-    public JavacOptionsResult buildTargetJavacOptions(JavacOptionsParams params) {
-        var project = projectProvider.get();
-        return bspMapper.buildTargetJavacOptions(project, params);
-    }
+  public JavacOptionsResult buildTargetJavacOptions(
+      CancelChecker cancelChecker, JavacOptionsParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.buildTargetJavacOptions(project, params);
+  }
 
-    public CppOptionsResult buildTargetCppOptions(CppOptionsParams params) {
-        var project = projectProvider.get();
-        return bspMapper.buildTargetCppOptions(project, params);
-    }
+  public CppOptionsResult buildTargetCppOptions(
+      CancelChecker cancelChecker, CppOptionsParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.buildTargetCppOptions(project, params);
+  }
 
-    public ScalacOptionsResult buildTargetScalacOptions(ScalacOptionsParams params) {
-        var project = projectProvider.get();
-        return bspMapper.buildTargetScalacOptions(project, params);
-    }
+  public ScalacOptionsResult buildTargetScalacOptions(
+      CancelChecker cancelChecker, ScalacOptionsParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.buildTargetScalacOptions(project, params);
+  }
 
-    public ScalaTestClassesResult buildTargetScalaTestClasses(ScalaTestClassesParams params) {
-        var project = projectProvider.get();
-        return bspMapper.buildTargetScalaTestClasses(project, params);
-    }
+  public ScalaTestClassesResult buildTargetScalaTestClasses(
+      CancelChecker cancelChecker, ScalaTestClassesParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.buildTargetScalaTestClasses(project, params);
+  }
 
-    public ScalaMainClassesResult buildTargetScalaMainClasses(ScalaMainClassesParams params) {
-        var project = projectProvider.get();
-        return bspMapper.buildTargetScalaMainClasses(project, params);
-    }
+  public ScalaMainClassesResult buildTargetScalaMainClasses(
+      CancelChecker cancelChecker, ScalaMainClassesParams params) {
+    var project = projectProvider.get(cancelChecker);
+    return bspMapper.buildTargetScalaMainClasses(project, params);
+  }
 
-    // TODO implement this endpoint to return libraries with maven coordinates that target depends on
-    // this should be helpful for 3rd party shared indexes in IntelliJ, however the endpoint is not
-    // yet used in the client
-    public DependencyModulesResult buildTargetDependencyModules(DependencyModulesParams params) {
-        return new DependencyModulesResult(Collections.emptyList());
-    }
+  // TODO implement this endpoint to return libraries with maven coordinates that target depends on
+  // this should be helpful for 3rd party shared indexes in IntelliJ, however the endpoint is not
+  // yet used in the client
+  public DependencyModulesResult buildTargetDependencyModules(
+      CancelChecker cancelChecker, DependencyModulesParams params) {
+    return new DependencyModulesResult(Collections.emptyList());
+  }
 
-    public RustWorkspaceResult rustWorkspace(RustWorkspaceParams params) {
-        var project = projectProvider.get();
-        return bspMapper.rustWorkspace(project, params);
+    public RustWorkspaceResult rustWorkspace(
+        CancelChecker cancelChecker, RustWorkspaceParams params) {
+      var project = projectProvider.get(cancelChecker);
+      return bspMapper.rustWorkspace(project, params);
     }
 }